tidy: Fix "multiplication implicit widening" in libmyth.
Fix the clang-tidy "implicit widening of multiplication result"
warnings.  This check is looking for places where a multiplication
operation may have overflowed the smaller type before the result is
converted to the bigger type. This will give erroneous results.

1) Cast multiplication arguments to a larger type.

2) Convert a variable to the appropriate type.

+++ mythtv/libs/libmyth/audio/audiooutputalsa.cpp
@@ -572,7 +572,7 @@ void AudioOutputALSA::WriteAudio(uchar *aubuf, int size)
                         .arg(lw * m_outputBytesPerFrame));
 
             frames -= lw;
-            tmpbuf += lw * m_outputBytesPerFrame; // bytes
+            tmpbuf += static_cast<ptrdiff_t>(lw) * m_outputBytesPerFrame; // bytes
             continue;
         }
